Move QgsMapRendererJob::needTemporaryImage to correct .cpp file
